refactorizacion y cambio de nombres
