+++ client/src/components/Layout/Header.vue
@@ -49,7 +49,6 @@ export default {
 }
 </script>
 
-<!-- Add "scoped" attribute to limit CSS to this component only -->
 <style scoped>
 nav{
   display: grid;
