+++ client/src/components/Layout/Login.vue
@@ -21,8 +21,7 @@ import { mapGetters, mapActions } from 'vuex'
 
 export default {
   data () {
-    return {
-    }
+    return {}
   },
   created () {
     this.checkLogged()
@@ -47,7 +46,6 @@ export default {
 }
 </script>
 
-<!-- Add "scoped" attribute to limit CSS to this component only -->
 <style scoped>
 .container{
   margin: 0;
