+++ client/src/filters.js
@@ -3,6 +3,10 @@ const filters =
   {
     name: 'snippet',
     execute: (value, long) => { return value.slice(0, long) + '...' }
+  },
+  {
+    name: 'capitalize',
+    execute: (value, long) => { return value.charAt(0).toUpperCase() }
   }
 ]
